requirements.txt
Flask==1.0.2
matplotlib==3.0.3
cognitive_face==1.5.0
requests==2.21.0
Pillow==6.1.0
gunicorn
